Update Cluster_doc.ecl
See ML-247 
In the ML\Tests\Benchmarks folder, the cluster_doc.ecl file included in the latest ML download generates syntax errors due to a typo in the ML.Cluster.KMeans near line 132. In addition, the DATASET referenced on Line 119 is a sprayed file that does not exist. If you comment out that line and uncomment the inline DATASET, and then change the lMatrix definition on line 14 to:

lMatrix:=ML.Types.NumericField;

The example will compile and execute correctly.

I have pushed the corrected code.